Allow the user to easy configure iodocs.
Now the configuration is stored under config/ directory that cointains:
 * config.json - the old config file
 * apiconfig.json - the specific api configuration file previously available under public/data/
 * apisites/ - a new directory wherein the user can store the configuration per each apisite

This improvement allows an easily configuration, improve the user configuration that can be done
using a symmbolic links outside the project itself.
A lot of github forks are used only to reconfigure the project and this skims new developer
to adopt iodocs.

Signed-off-by: Claudio Mignanti <c.mignanti@gmail.com>